feat(workflows): "Write with AI" call-to-action on an empty body
