+++ packages/nc-gui/components/nc/form-builder/input/WorkflowInputTools.vue
@@ -1,8 +1,7 @@
 <script setup lang="ts">
 import type { Editor } from '@tiptap/vue-3'
-import { DOMSerializer } from '@tiptap/pm/model'
-import type { DocAiImproveMode, VariableDefinition } from 'nocodb-sdk'
-import { useWorkflowEmailAi } from '~/composables/useWorkflowEmailAi'
+import type { VariableDefinition } from 'nocodb-sdk'
+import { useWorkflowEmailAi } from '#imports'
 import { EMAIL_FONTS } from '~/helpers/tiptap-markdown/extensions/marks/fontFamily'
 import { EMAIL_FONT_SIZES } from '~/helpers/tiptap-markdown/extensions/marks/fontSize'
 import { HIGHLIGHT_COLORS } from '~/helpers/tiptap-markdown/extensions/marks/highlight'
@@ -29,82 +28,12 @@ const emits = defineEmits<{
   (e: 'aiResult', payload: { html: string; mode: 'write' | 'rewrite' }): void
 }>()
 
-const { $e } = useNuxtApp()
+const { available: aiAvailable } = useWorkflowEmailAi()
 
-const { available: aiAvailable, loading: aiLoading, aiWrite, aiRewrite, abort: aiAbort } = useWorkflowEmailAi()
-
-const aiOpen = ref(false)
-
-const aiInstruction = ref('')
-
-const aiInputRef = ref<HTMLTextAreaElement>()
-
-const hasSelection = computed(() => !props.editor.state.selection.empty)
-
-const rewriteModes: { mode: DocAiImproveMode; label: string }[] = [
-  { mode: 'writing', label: 'labels.docAiImproveWriting' },
-  { mode: 'grammar', label: 'labels.docAiFixGrammar' },
-  { mode: 'shorter', label: 'labels.docAiMakeShorter' },
-  { mode: 'longer', label: 'labels.docAiMakeLonger' },
-]
-
-const toneModes: DocAiImproveMode[] = ['professional', 'friendly', 'casual', 'confident']
-
-// Flatten the upstream variable tree into what the prompt lists (leaf keys only).
-const aiVariables = computed(() => {
-  const out: { key: string; name: string; type?: string }[] = []
-  const walk = (vars: any[], prefix = '') => {
-    for (const v of vars || []) {
-      const name = prefix ? `${prefix} › ${v.name}` : v.name
-      if (v.children?.length) walk(v.children, name)
-      else if (v.key && !String(v.key).includes('.map(')) out.push({ key: v.key, name, type: v.type })
-    }
-  }
-  walk(props.variables)
-  return out.slice(0, 60)
-})
-
-watch(aiOpen, (open) => {
-  if (open) nextTick(() => aiInputRef.value?.focus())
-  else aiAbort()
-})
-
-async function runAiWrite() {
-  const instruction = aiInstruction.value.trim()
-  if (!instruction || aiLoading.value) return
-  $e('a:workflow:email:ai:write')
-  const html = await aiWrite({
-    instruction,
-    currentBody: props.editor.isEmpty ? '' : props.editor.getHTML(),
-    variables: aiVariables.value,
-  })
-  if (!html) return
-  emits('aiResult', { html, mode: 'write' })
-  aiInstruction.value = ''
-  aiOpen.value = false
-}
-
-// Selected content as HTML, with variable chips sent as their {{ }} tokens so the model
-// sees (and preserves) the real expression rather than a display label.
-function selectionAsHtml(): string {
-  const { from, to } = props.editor.state.selection
-  const slice = props.editor.state.doc.slice(from, to)
-  const container = document.createElement('div')
-  container.appendChild(DOMSerializer.fromSchema(props.editor.schema).serializeFragment(slice.content))
-  container.querySelectorAll('span[data-type="workflowExpression"]').forEach((el) => {
-    el.replaceWith(document.createTextNode(el.getAttribute('data-expression') || ''))
-  })
-  return container.innerHTML
-}
-
-async function runAiRewrite(mode: DocAiImproveMode) {
-  if (aiLoading.value || props.editor.state.selection.empty) return
-  $e('a:workflow:email:ai:rewrite', { mode })
-  const html = await aiRewrite({ html: selectionAsHtml(), mode, variables: aiVariables.value })
-  if (!html) return
-  emits('aiResult', { html, mode: 'rewrite' })
-  aiOpen.value = false
-}
+// A group whose only tool is unavailable must not leave a stray divider behind.
+const visibleGroups = computed(() =>
+  props.groups.map((group) => group.filter((tool) => tool.type !== 'ai' || aiAvailable.value)).filter((group) => group.length),
+)
 
 // The email body's ink. Highlight pastels are picked against it, so the preview must not
 // fall back to theme text, which goes light in dark mode and vanishes on the swatch.
@@ -199,7 +128,7 @@ function applyHighlight(color: string) {
 </script>
 
 <template>
-  <template v-for="(group, gi) in groups" :key="gi">
+  <template v-for="(group, gi) in visibleGroups" :key="gi">
     <div v-if="gi > 0" class="nc-email-format-divider" />
 
     <template v-for="tool in group" :key="tool.key">
@@ -339,84 +268,31 @@ function applyHighlight(color: string) {
         </template>
       </NcDropdown>
 
-      <NcDropdown
-        v-else-if="tool.type === 'ai' && aiAvailable"
-        v-model:visible="aiOpen"
-        placement="bottomLeft"
-        :overlay-style="{ zIndex: 10002 }"
+      <NcFormBuilderInputWorkflowInputAi
+        v-else-if="tool.type === 'ai'"
+        :editor="editor"
+        :variables="variables"
+        @result="emits('aiResult', $event)"
       >
-        <NcTooltip :title="$t('labels.writeWithAi')">
-          <NcButton
-            size="xs"
-            type="text"
-            class="nc-workflow-format-btn nc-email-ai-btn"
-            :class="{ 'is-active': aiOpen || aiLoading }"
-            :loading="aiLoading"
-            data-testid="nc-workflow-richtext-ai-btn"
-            @mousedown.prevent
-            @click.stop="aiOpen = !aiOpen"
-          >
-            <GeneralIcon v-if="!aiLoading" icon="ncAutoAwesome" class="w-4 h-4 text-nc-content-brand" />
-          </NcButton>
-        </NcTooltip>
-
-        <template #overlay>
-          <div class="nc-email-ai-menu" @mousedown.stop @click.stop>
-            <div class="nc-email-ai-title">
-              <GeneralIcon icon="ncAutoAwesome" class="w-3.5 h-3.5 text-nc-content-brand" />
-              {{ $t('labels.writeWithAi') }}
-            </div>
-            <textarea
-              ref="aiInputRef"
-              v-model="aiInstruction"
-              class="nc-email-ai-input"
-              rows="3"
-              :placeholder="$t('placeholder.describeEmail')"
-              data-testid="nc-workflow-richtext-ai-input"
-              @keydown.enter.exact.prevent="runAiWrite"
-              @keydown.esc.stop.prevent="aiOpen = false"
-            />
-            <div class="flex items-center justify-between gap-2">
-              <span v-if="aiVariables.length" class="text-tiny text-nc-content-gray-muted">
-                {{ aiVariables.length }} {{ $t('general.variables').toLowerCase() }}
-              </span>
-              <span v-else />
-              <NcButton
-                size="xs"
-                type="primary"
-                :disabled="!aiInstruction.trim() || aiLoading"
-                :loading="aiLoading"
-                data-testid="nc-workflow-richtext-ai-generate"
-                @click="runAiWrite"
-              >
-                {{ $t('general.generate') }}
-              </NcButton>
-            </div>
-
-            <template v-if="hasSelection">
-              <div class="nc-email-ai-divider" />
-              <div class="nc-email-ai-section">{{ $t('labels.rewriteSelection') }}</div>
-              <button
-                v-for="m in rewriteModes"
-                :key="m.mode"
-                class="nc-email-ai-item"
-                :disabled="aiLoading"
-                @click="runAiRewrite(m.mode)"
-              >
-                {{ $t(m.label) }}
-              </button>
-              <div class="nc-email-ai-section">{{ $t('labels.aiTone') }}</div>
-              <div class="flex flex-wrap gap-1 px-1 pb-1">
-                <button v-for="m in toneModes" :key="m" class="nc-email-ai-chip" :disabled="aiLoading" @click="runAiRewrite(m)">
-                  {{ m }}
-                </button>
-              </div>
-            </template>
-          </div>
+        <template #default="{ open, loading, toggle }">
+          <NcTooltip :title="$t('labels.writeWithAi')">
+            <NcButton
+              size="xs"
+              type="text"
+              class="nc-workflow-format-btn nc-email-ai-btn"
+              :class="{ 'is-active': open || loading }"
+              :loading="loading"
+              data-testid="nc-workflow-richtext-ai-btn"
+              @mousedown.prevent
+              @click.stop="toggle"
+            >
+              <GeneralIcon v-if="!loading" icon="ncAutoAwesome" class="w-4 h-4 text-nc-content-brand" />
+            </NcButton>
+          </NcTooltip>
         </template>
-      </NcDropdown>
+      </NcFormBuilderInputWorkflowInputAi>
 
-      <NcTooltip v-else-if="tool.type !== 'ai'" :title="$t(tool.label!)">
+      <NcTooltip v-else :title="$t(tool.label!)">
         <NcButton
           size="xs"
           type="text"
@@ -457,54 +333,6 @@ function applyHighlight(color: string) {
 }
 
 // Overlays render in body, so these stay unscoped.
-.nc-email-ai-menu {
-  @apply flex flex-col gap-2 p-3 rounded-lg bg-nc-bg-default border-1 border-nc-border-gray-medium;
-  width: 320px;
-  box-shadow: 0 8px 24px rgba(16, 16, 21, 0.12);
-
-  .nc-email-ai-title {
-    @apply flex items-center gap-1.5 text-small font-semibold text-nc-content-gray;
-  }
-
-  .nc-email-ai-input {
-    @apply w-full px-2.5 py-2 text-small rounded-md border-1 border-nc-border-gray-medium outline-none resize-none;
-    @apply focus:border-nc-border-brand;
-    line-height: 1.5;
-  }
-
-  .nc-email-ai-divider {
-    @apply h-px bg-nc-border-gray-light my-1;
-  }
-
-  .nc-email-ai-section {
-    @apply text-tiny font-semibold uppercase text-nc-content-gray-muted px-1;
-    letter-spacing: 0.04em;
-  }
-
-  .nc-email-ai-item {
-    @apply flex items-center h-7.5 px-2 rounded-md cursor-pointer text-left bg-transparent border-0 text-nc-content-gray;
-    font-size: 13px;
-
-    &:hover {
-      @apply bg-nc-bg-gray-light;
-    }
-
-    &:disabled {
-      @apply opacity-50 cursor-default;
-    }
-  }
-
-  .nc-email-ai-chip {
-    @apply h-6.5 px-2 rounded-md cursor-pointer capitalize text-nc-content-gray-subtle;
-    @apply border-1 border-nc-border-gray-medium bg-nc-bg-default;
-    font-size: 12px;
-
-    &:hover {
-      @apply border-nc-border-brand bg-nc-bg-brand text-nc-content-brand;
-    }
-  }
-}
-
 .nc-email-typo-picker {
   // NcDropdown's overlay already draws the border, radius and shadow; repeating them here is
   // what doubled the outline.
